refactor Nillion snap client logic to useNillionSnapClient React hook
